rename to make pictogram creation/login share bloc
